Split executable and library for obelisk-asset-manifest
unix-compat, a library dependency of obelisk-asset-manifest does not build with the new js backend.

This is a problem because parts of obelisk-asset-manifest are being used in js backend TH.

However unix-compat is not needed for that TH, thus we move all code which depends on unix-compat into the executable, which only is needed with the native backend.